requirements.txt
Flask==1.0.2
requests==2.18.4
gunicorn==19.8.1
django-htmlmin==0.10.0


